Remove AppConfigSpec/AppConfig distriction.
Thanks to the hand-waving of no-type-checks-at-runtime, we can use
the same AppConfig class to both hold our initial ConfigOption
instances as well as the post-evaluation values themselves.

This requires some casting/blatant lying to the type system, but
the end result is pretty slick.

In particular it means "find usages" on the AppConfig fields actually
works, where as previously the AppConfigSpec fields were unreferenced
and so always looked unused.