Clean up ConfigContext and ignoreErrors.
Specifically remove ConfigContext and just go with an Environment
type as the 2nd argument. ConfigContext was meant to be an area
for API flexibility, e.g. to have other non-env ways of doing
lookups, but the "name -> value" of Environment is already pretty
generic, and the 3rd arg of ConfigOptions provides the API
flexibility.

Speaking of the 3rd parameter, it was a boolean ignoreErrors, but
that doesn't provide flexibility for adding more down the road,
so move to a standard options hash idiom.